requirements.txt
absl-py==0.12.0
appdirs==1.4.4
astroid==2.5.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
daemonize==2.5.0
distlib==0.3.1
filelock==3.0.12
google-auth==1.28.0
google-auth-oauthlib==0.4.3
grpcio==1.36.1
guildai==0.7.3
idna==2.10
isort==5.7.0
Jinja2==2.11.3
joblib==1.0.1
lazy-object-proxy==1.5.2
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
pkginfo==1.7.0
protobuf==3.15.6
psutil==5.8.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.7.2
pyparsing==2.4.7
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.43
scikit-learn==0.24.1
scikit-optimize==0.8.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
tabview==1.4.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
tokenizers==0.10.1
toml==0.10.2
torch==1.8.0
tqdm==4.59.0
transformers==4.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.3
Werkzeug==1.0.1
wrapt==1.12.1
